feat: Improved Ikkegol protocol implementation
The config packet can be up to 40 bytes in size depending on what is configured.
The largest part of this is the string mode which can enter up to 38 characters.

The first byte of the packet has been identified as the number of bytes in the message.
In most cases it is rounded up to nearest multiple of 8, but in some config types it shows the exact number of bytes.
The number of bytes to request is always the nearest upper multiple of this size field.

The reading routine has been updated to read in 8 byte blocks up to a maximum of 40 bytes depending on the size field.