feat: Refactored CLI and implemented remaining command line
Was not happy with the args library. It did not allow the interface wanted.
Many libraries were looked at, but none seemed to allow nested "commands" (for the set command)

A custom implementation was chosen which provided the exact interface wanted.
As a plus, the help screens are more useful than what those libraries usually generate.

Implemented all the remaining set commands.
Extracted out the printConfig function from the show command so it can be used in the set command too.