feat: Now able to update trigger modes
Now pedals can be configured respond to release instead of press.

NOTE: Changing this mode often causes the pedal to continuously fire all pedals. It does this in the official software too. The fix is to unplug and replug it. Will have to see if we can reboot the device or something